#battery #bugfix Bug 1: Was not outputting annual grid export/import, modified where the calculation occurred Bug 2: The target dispatch was being affected by the capacity being adjusted by the temperature.  Adjusted the current derate for temperature

+++ shared/lib_battery.h
@@ -28,7 +28,7 @@ public:
 
 	void add(std::string message);
 	int total_message_count();
-	int message_count(int index);
+	size_t message_count(int index);
 	std::string get_message(int index);
 	std::string construct_log_count_string(int index);
